Start migration to SS4
